packages/libs/medflowai/requirements.txt
openai>=1.78.0,<2.0.0
openai-agents>=0.0.14,<0.1.0
google-genai>=1.14.0,<1.15.0
pydantic-ai>=0.1.11,<0.2.0
